In progress
This is a document management webapp where users can manage documents in the following format (pdf, xls, csv, doc) and it includes an e-signature feature where users can include their signature to any of the document and send to a recipient.

It was built using html, css, javascript and php. It includes a simple restful API (docusign) which enables the embedded signature.